primario y confirm added

+++ sectionPages/tuEspacio/Fotos.jsx
@@ -18,6 +18,7 @@ import 'swiper/css/free-mode';
 import 'swiper/css/pagination';
 
 import style from './Fotos.module.css';
+import { Dropdown, DropdownButton } from 'react-bootstrap';
 
 export const Fotos = ({ sobre, limite }) => {
   const { data: session, update: sessionUpdate, status } = useSession();
@@ -60,7 +61,7 @@ export const Fotos = ({ sobre, limite }) => {
   const deleteTest = async (id, publicId) => {
     // e.preventDefault();
     if (publicId) {
-      console.log(publicId);
+      if (!confirm('Estas seguro que deseas eliminarlo?')) return;
       try {
         await axios.delete(
           `${process.env.BASE_URL}/espacio/publicar?publicId=${publicId}&sobre=${sobre}`,
@@ -82,13 +83,29 @@ export const Fotos = ({ sobre, limite }) => {
   };
 
   // Coloca una imagen al inicio convirtiendolo en primaria
-  const onPrimaryImg = (id) => {
-    // Elimina primero la imagen, luego guarda en un nuevo array
-    const changePlace = gallery.filter((item, idx) => idx !== id);
-    // Agarra la imagen que queremos poner de primario desde gallery y lo mete en el nuevo array
-    changePlace.unshift(gallery[id]);
-    // Metiendo el cambio a gallery
-    setGallery(changePlace);
+  const onPrimaryImg = async (id) => {
+    try {
+      if (confirm('Seguro que quieres convertirlo a primario?')) {
+        // Elimina primero la imagen, luego guarda en un nuevo array
+        const changePlace = gallery.filter((item, idx) => idx !== id);
+        // Agarra la imagen que queremos poner de primario desde gallery y lo mete en el nuevo array
+        changePlace.unshift(gallery[id]);
+
+        const form = {
+          [`${sobre}Img`]: changePlace,
+        };
+        const { data } = await axios.put(
+          `${process.env.BASE_URL}/espacio/publicar`,
+          form,
+          config
+        );
+        console.log(data[`${sobre}Img`]);
+        // Metiendo el cambio a gallery
+        setGallery(data[`${sobre}Img`]);
+      }
+    } catch (error) {
+      toast.error(getError(error));
+    }
   };
 
   const testing = async () => {
@@ -197,7 +214,10 @@ export const Fotos = ({ sobre, limite }) => {
         Imagenes subida en la nube {gallery.length} de {limite}
       </h5>
 
-      <div className="w-100">
+      <div
+        className="w-100 bg-secondary bg-opacity-50 py-2"
+        style={{ border: '5px inset', overflowX: 'hidden' }}
+      >
         <Swiper
           style={{
             '--swiper-navigation-color': 'red',
@@ -206,7 +226,7 @@ export const Fotos = ({ sobre, limite }) => {
             padding: '0 10px',
           }}
           spaceBetween={10}
-          slidesPerView={4}
+          slidesPerView={3}
           freeMode={true}
           // navigation={true}
           pagination={true}
@@ -215,20 +235,32 @@ export const Fotos = ({ sobre, limite }) => {
         >
           {gallery.map(({ url, publicId }, idx) => (
             <SwiperSlide key={publicId}>
-              <div className={`${style.imgBox} position-relative h-100`}>
+              <div className={`${style.gallery} position-relative h-100`}>
                 <img
                   src={url}
                   decoding="async"
                   className="h-100"
-                  onClick={() => onPrimaryImg(idx, publicId)}
+                  // onClick={() => onPrimaryImg(idx, publicId)}
                 />
-                <button
+                {/* <button
                   className="position-absolute btn btn-outline-danger"
                   type="button"
                   onClick={() => deleteTest(idx, publicId)}
                 >
                   <i className="bi bi-x-lg"></i>
-                </button>
+                </button> */}
+                <DropdownButton
+                  className={`${style.drop}`}
+                  title={<i class="bi bi-three-dots-vertical"></i>}
+                  drop={'start'}
+                >
+                  <Dropdown.Item onClick={() => deleteTest(idx, publicId)}>
+                    Eliminar Imagen
+                  </Dropdown.Item>
+                  <Dropdown.Item onClick={() => onPrimaryImg(idx, publicId)}>
+                    Convertir a Primario
+                  </Dropdown.Item>
+                </DropdownButton>
               </div>
             </SwiperSlide>
           ))}
